pyproject.toml
[tool.poetry]
name = "tfg-dietplanner"
version = "0.1.0"
description = ""
authors = ["Alejandro Olivares <a.olivaresdelrey.pierres@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.6"
invoke = "^2.1.3"
pytest = "^7.4.3"
python-decouple = "^3.8"
sqlalchemy = "^2.0.23"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
